Move seccomp helper methods and tests to platform-specific files.
