+++ pkg/kubelet/dockershim/BUILD
@@ -82,6 +82,7 @@ go_test(
         "docker_legacy_test.go",
         "docker_sandbox_test.go",
         "docker_service_test.go",
+        "helpers_linux_test.go",
         "helpers_test.go",
         "naming_test.go",
         "security_context_test.go",
